+++ src/parser/enrichers/class/monk/DeflectAttacks.mjs
@@ -6,7 +6,16 @@ export default class DeflectAttacks extends DDBEnricherData {
   get additionalActivities() {
     return [
       { action: { name: "Deflect Attack", type: "class" } },
-      { action: { name: "Deflect Attack: Redirect Attack", type: "class" } },
+      {
+        action: { name: "Deflect Attack: Redirect Attack", type: "class" },
+        overrides: {
+          data: {
+            damage: {
+              onSave: "none",
+            },
+          },
+        },
+      },
     ];
   }